+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Redundant ghost checkpoint at bottom of page is removed (@ffiedrich)
 - After dragging a checkpoint, the menu disappears unless it is hovered (@ffiedrich)
 - Path toggle listeners are added to hidden skills on first reveal (@ffiedrich)
+- Completion status of revealed hidden skills is updated without page reload (@ffiedrich)
+- Revealed skills are not hidden again when required tasks are marked as uncomplete (@rglans)
 
 ## [2.2.6]